--- evacuees/src/UI/embc-evacuee/src/app/core/services/api/api-configuration.ts
@@ -1,19 +1,0 @@
-/* tslint:disable */
-import { Injectable } from '@angular/core';
-
-/**
- * Global configuration
- */
-@Injectable({
-  providedIn: 'root',
-})
-export class ApiConfiguration {
-  rootUrl: string = 'http://localhost:2010';
-}
-
-/**
- * Parameters for `ApiModule.forRoot()`
- */
-export interface ApiConfigurationParams {
-  rootUrl?: string;
-}
