--- evacuees/src/UI/embc-evacuee/src/app/core/services/api/api.module.ts
@@ -1,49 +1,0 @@
-/* tslint:disable */
-import { NgModule, ModuleWithProviders, SkipSelf, Optional } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { ApiConfiguration, ApiConfigurationParams } from './api-configuration';
-
-import { ErrorService } from './services/error.service';
-import { RegistrationService } from './services/registration.service';
-import { LocationService } from './services/location.service';
-
-/**
- * Module that provides all services and configuration.
- */
-@NgModule({
-  imports: [],
-  exports: [],
-  declarations: [],
-  providers: [
-    ErrorService,
-    RegistrationService,
-    LocationService,
-    ApiConfiguration
-  ],
-})
-export class ApiModule {
-  static forRoot(params: ApiConfigurationParams): ModuleWithProviders<ApiModule> {
-    return {
-      ngModule: ApiModule,
-      providers: [
-        {
-          provide: ApiConfiguration,
-          useValue: params
-        }
-      ]
-    }
-  }
-
-  constructor( 
-    @Optional() @SkipSelf() parentModule: ApiModule,
-    @Optional() http: HttpClient
-  ) {
-    if (parentModule) {
-      throw new Error('ApiModule is already loaded. Import in your base AppModule only.');
-    }
-    if (!http) {
-      throw new Error('You need to import the HttpClientModule in your AppModule! \n' +
-      'See also https://github.com/angular/angular/issues/20575');
-    }
-  }
-}
