--- evacuees/src/UI/embc-evacuee/src/app/core/services/api/models.ts
@@ -1,14 +1,0 @@
-export { RegistrationResult } from './models/registration-result';
-export { ProblemDetails } from './models/problem-details';
-export { AnonymousRegistration } from './models/anonymous-registration';
-export { Registration } from './models/registration';
-export { PersonDetails } from './models/person-details';
-export { ContactDetails } from './models/contact-details';
-export { Address } from './models/address';
-export { NeedsAssessment } from './models/needs-assessment';
-export { InsuranceOption } from './models/insurance-option';
-export { Pet } from './models/pet';
-export { Country } from './models/country';
-export { StateProvince } from './models/state-province';
-export { Jurisdiction } from './models/jurisdiction';
-export { JurisdictionType } from './models/jurisdiction-type';
