--- evacuees/src/UI/embc-evacuee/src/app/core/services/api/services.ts
@@ -1,3 +1,0 @@
-export { ErrorService } from './services/error.service';
-export { RegistrationService } from './services/registration.service';
-export { LocationService } from './services/location.service';
